+++ sdk/resourcemanager/streamanalytics/armstreamanalytics/ze_generated_example_clusters_client_test.go
@@ -154,7 +154,7 @@ func ExampleClustersClient_BeginDelete() {
 }
 
 // Generated from example definition: https://github.com/Azure/azure-rest-api-specs/tree/main/specification/streamanalytics/resource-manager/Microsoft.StreamAnalytics/stable/2020-03-01/examples/Cluster_ListBySubscription.json
-func ExampleClustersClient_ListBySubscription() {
+func ExampleClustersClient_NewListBySubscriptionPager() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
 		log.Fatalf("failed to obtain a credential: %v", err)
@@ -166,7 +166,7 @@ func ExampleClustersClient_ListBySubscription() {
 		log.Fatalf("failed to create client: %v", err)
 		return
 	}
-	pager := client.ListBySubscription(nil)
+	pager := client.NewListBySubscriptionPager(nil)
 	for pager.More() {
 		nextResult, err := pager.NextPage(ctx)
 		if err != nil {
@@ -181,7 +181,7 @@ func ExampleClustersClient_ListBySubscription() {
 }
 
 // Generated from example definition: https://github.com/Azure/azure-rest-api-specs/tree/main/specification/streamanalytics/resource-manager/Microsoft.StreamAnalytics/stable/2020-03-01/examples/Cluster_ListByResourceGroup.json
-func ExampleClustersClient_ListByResourceGroup() {
+func ExampleClustersClient_NewListByResourceGroupPager() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
 		log.Fatalf("failed to obtain a credential: %v", err)
@@ -193,7 +193,7 @@ func ExampleClustersClient_ListByResourceGroup() {
 		log.Fatalf("failed to create client: %v", err)
 		return
 	}
-	pager := client.ListByResourceGroup("<resource-group-name>",
+	pager := client.NewListByResourceGroupPager("<resource-group-name>",
 		nil)
 	for pager.More() {
 		nextResult, err := pager.NextPage(ctx)
@@ -209,7 +209,7 @@ func ExampleClustersClient_ListByResourceGroup() {
 }
 
 // Generated from example definition: https://github.com/Azure/azure-rest-api-specs/tree/main/specification/streamanalytics/resource-manager/Microsoft.StreamAnalytics/stable/2020-03-01/examples/Cluster_ListStreamingJobs.json
-func ExampleClustersClient_ListStreamingJobs() {
+func ExampleClustersClient_NewListStreamingJobsPager() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
 		log.Fatalf("failed to obtain a credential: %v", err)
@@ -221,7 +221,7 @@ func ExampleClustersClient_ListStreamingJobs() {
 		log.Fatalf("failed to create client: %v", err)
 		return
 	}
-	pager := client.ListStreamingJobs("<resource-group-name>",
+	pager := client.NewListStreamingJobsPager("<resource-group-name>",
 		"<cluster-name>",
 		nil)
 	for pager.More() {