+++ sdk/resourcemanager/streamanalytics/armstreamanalytics/ze_generated_example_functions_client_test.go
@@ -157,7 +157,7 @@ func ExampleFunctionsClient_Get() {
 }
 
 // Generated from example definition: https://github.com/Azure/azure-rest-api-specs/tree/main/specification/streamanalytics/resource-manager/Microsoft.StreamAnalytics/stable/2020-03-01/examples/Function_ListByStreamingJob.json
-func ExampleFunctionsClient_ListByStreamingJob() {
+func ExampleFunctionsClient_NewListByStreamingJobPager() {
 	cred, err := azidentity.NewDefaultAzureCredential(nil)
 	if err != nil {
 		log.Fatalf("failed to obtain a credential: %v", err)
@@ -169,7 +169,7 @@ func ExampleFunctionsClient_ListByStreamingJob() {
 		log.Fatalf("failed to create client: %v", err)
 		return
 	}
-	pager := client.ListByStreamingJob("<resource-group-name>",
+	pager := client.NewListByStreamingJobPager("<resource-group-name>",
 		"<job-name>",
 		&armstreamanalytics.FunctionsClientListByStreamingJobOptions{Select: nil})
 	for pager.More() {