+++ sdk/resourcemanager/streamanalytics/armstreamanalytics/zz_generated_constants.go
@@ -10,7 +10,7 @@ package armstreamanalytics
 
 const (
 	moduleName    = "armstreamanalytics"
-	moduleVersion = "v0.4.0"
+	moduleVersion = "v0.5.0"
 )
 
 // AuthenticationMode - Authentication Mode. Valid modes are ConnectionString, Msi and 'UserToken'.